+++ CHANGELOG.md
@@ -11,6 +11,7 @@ This project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html
 - Improved documentation; README now includes command options.
 
 ### Breaking changes
+- Removed deprecated executables
 
 
 ## 3.2.0 2020-07-15