+++ CHANGELOG.md
@@ -4,8 +4,12 @@ All notable changes to this project will be documented in this file.
 
 This project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-
 ## Unreleased
+### Compatible changes
+### Breaking changes
+
+
+## 4.0.0 2020-07-30
 
 ### Compatible changes
 - Improved documentation; README now includes command options.